requirements.txt
langchain 
wikipedia 
langchain_community 
langchain_openai
langchain_anthropic
python_dotenv
pydantic
duckduckgo_search